move share-link into tab instead of modal
